+++ rust/ommx/src/random.rs
@@ -40,6 +40,7 @@ mod parameter;
 mod parametric_instance;
 mod polynomial;
 mod quadratic;
+mod samples;
 mod state;
 
 pub use constraint::*;
@@ -47,6 +48,7 @@ pub use decision_variable::*;
 pub use function::*;
 pub use instance::*;
 pub use parameter::*;
+pub use samples::*;
 pub use state::*;
 
 pub use proptest::test_runner::TestRunner as Rng;
